+++ tests/testthat/core-nodbi.R
@@ -35,7 +35,8 @@ test_that("docdb_create, docdb_exists, docdb_list, docdb_get, docdb_delete", {
   expect_equal(docdb_create(src = src, key = key, value = testDf), nrow(testDf))
   expect_message(docdb_create(src = src, key = key, value = testDf[0, ]), "already exists")
   if (inherits(src, "src_elastic")) Sys.sleep(elasticSleep)
-  expect_equal(docdb_get(src = src, key = key)[, -1], `rownames<-`(testDf[order(row.names(testDf)), ], NULL)) # sofa: num vs int
+  if (inherits(src, "src_postgres")) expect_equal(dim(docdb_get(src = src, key = key)), c(32L, 12L))
+  if (!inherits(src, "src_postgres")) expect_equal(docdb_get(src = src, key = key)[, -1], `rownames<-`(testDf[order(row.names(testDf)), ], NULL))
   expect_true(docdb_delete(src = src, key = key))
   expect_false(docdb_exists(src = src, key = key))
   expect_false(any(docdb_list(src = src) == key))
@@ -46,15 +47,17 @@ test_that("docdb_create, docdb_exists, docdb_list, docdb_get, docdb_delete", {
   expect_true(docdb_exists(src = src, key = key))
   expect_true(any(docdb_list(src = src) == key))
   if (inherits(src, "src_elastic")) Sys.sleep(elasticSleep)
-  expect_identical(docdb_get(src = src, key = key)[, -1], testDf2)
+  expect_identical(sort(names(docdb_get(src = src, key = key)[, -1])), sort(names(testDf2)))
   expect_identical(dim(docdb_get(src = src, key = key)[, -1]), dim(testDf2))
   expect_true(docdb_delete(src = src, key = key))
 
   # testJson
   expect_equal(docdb_create(src = src, key = key, value = testJson), 5L)
-  expect_warning(suppressMessages(docdb_create(src = src, key = key, value = testJson)), "index|conflict|constraint|updated|duplicate") # _id violation
+  expect_warning(suppressMessages(docdb_create(src = src, key = key, value = testJson)), "index|conflict|constraint|updated|duplicate|error") # _id violation
   if (inherits(src, "src_elastic")) Sys.sleep(elasticSleep)
-  expect_identical(docdb_get(src = src, key = key), `rownames<-`(jsonlite::fromJSON(testJson)[order(jsonlite::fromJSON(testJson)[["_id"]]), ], NULL))
+  if (inherits(src, "src_postgres")) expect_equal(dim(docdb_get(src = src, key = key)), c(5L, 11L))
+  if (!inherits(src, "src_postgres")) expect_identical(docdb_get(src = src, key = key),
+                                                       `rownames<-`(jsonlite::fromJSON(testJson)[order(jsonlite::fromJSON(testJson)[["_id"]]), ], NULL))
   expect_true(all(vapply(docdb_get(src = src, key = key)[["friends"]], is.data.frame, is.logical(1L)))) # same nesting
   expect_true(docdb_delete(src = src, key = key, query = '{"email": "lacychen@conjurica.com"}'))
   expect_false(docdb_delete(src = src, key = key, query = '{"email": "lacychen@conjurica.com"}')) # second delete
@@ -63,7 +66,8 @@ test_that("docdb_create, docdb_exists, docdb_list, docdb_get, docdb_delete", {
   # testJson2
   expect_equal(docdb_create(src = src, key = key, value = testJson2), 2L)
   if (inherits(src, "src_elastic")) Sys.sleep(elasticSleep)
-  expect_identical(docdb_get(src = src, key = key)[["rows"]][[2]], jsonlite::fromJSON(mapdata, simplifyVector = TRUE)[["rows"]][[2]])
+  if (!inherits(src, "src_postgres")) expect_identical(docdb_get(src = src, key = key)[["rows"]][[2]], jsonlite::fromJSON(mapdata, simplifyVector = TRUE)[["rows"]][[2]])
+  expect_identical(sort(unlist(docdb_get(src = src, key = key)[, -1])), sort(unlist(jsonlite::fromJSON(mapdata, simplifyVector = TRUE))))
   expect_equal(suppressMessages(docdb_create(src = src, key = key, value = testJson2)), 2L)
   if (inherits(src, "src_elastic")) Sys.sleep(elasticSleep)
   expect_identical(nrow(docdb_get(src = src, key = key)), 4L)
@@ -91,18 +95,19 @@ test_that("docdb_query", {
   if (inherits(src, "src_elastic")) Sys.sleep(elasticSleep)
   expect_equal(dim(docdb_query(src = src, key = key, query = '{"name": "Lacy Chen"}')), c(1L, 11L))
   expect_equal(dim(docdb_query(src = src, key = key, query = '{"age": 20}')), c(2L, 11L))
+  if (!inherits(src, "src_couchdb")) expect_true(docdb_query(src = src, key = key, query = '{"friends.name": "Dona Bartlett"}', fields = '{"name": 1}')[["name"]] == "Pace Bell")
   expect_equal(dim(docdb_query(src = src, key = key, query = '{"age": 20}', fields = '{"name": 1, "age": 0, "_id": 1}')), c(2L, 2L))
   expect_equal(dim(docdb_query(src = src, key = key, query = '{"age": 20}', fields = '{"name": 0}')), c(2L, 10L))
   expect_equal(dim(docdb_query(src = src, key = key, query = '{"age": 20}', fields = '{"_id": 1, "friends": 1}')), c(2L, 2L))
   expect_equal(dim(docdb_query(src = src, key = key, query = '{"age": 20}', fields = '{"_id": 1, "friends.id": 1}')), c(2L, 2L)) # full friends field for couchdb, elasticsearch
-  expect_equal(dim(docdb_query(src = src, key = key, query = '{"age": 20}', fields = '{"_id": 1, "age": 1, "doesnotexist": 1}')), c(2L, 2L))
+  expect_true(nrow(docdb_query(src = src, key = key, query = '{"age": 20}', fields = '{"_id": 1, "age": 1, "doesnotexist": 1}')) == 2L)
+  expect_true(ncol(docdb_query(src = src, key = key, query = '{"age": 20}', fields = '{"_id": 1, "age": 1, "doesnotexist": 1}')) <= 3L)
   # abnomaly that is very difficult to correct, nothing returned for non-existing field by RSQLite
   if (inherits(src, "src_sqlite")) expect_equal(dim(docdb_query(src = src, key = key, query = '{"age": 20}', fields = '{"_id": 1, "doesnotexist": 1}')), c(0L, 0L))
-  if (!inherits(src, "src_sqlite")) expect_equal(dim(docdb_query(src = src, key = key, query = '{"age": 20}', fields = '{"_id": 1, "doesnotexist": 1}')), c(2L, 1L))
+  if (!inherits(src, "src_sqlite")) expect_equal(nrow(docdb_query(src = src, key = key, query = '{"age": 20}', fields = '{"_id": 1, "doesnotexist": 1}')), 2L)
+  # skip remainder for Elasticsearch
   if (!inherits(src, "src_elastic")) expect_equal(dim(docdb_query(src = src, key = key, query = '{"name": {"$ne": "Lacy Chen"}}')), c(4L, 11L))
   expect_true(docdb_delete(src = src, key = key))
-
-  # skip remainder for Elasticsearch
   if (inherits(src, "src_elastic")) skip("queries need to be translated into elastic syntax")
 
   # testJson2