+++ tests/testthat/helper-nodbi.R
@@ -40,3 +40,12 @@ skip_if_no_elastic <- function() {
     skip("elasticsearch is not available")
   }
 }
+
+skip_if_no_postgres <- function() {
+  testthat::skip_if_not_installed("RPostgres")
+  if (inherits(try(tmp <- src_postgres(), silent = TRUE), "try-error")) {
+    skip("postgres is not available")
+  }
+  RPostgres::dbDisconnect(tmp$con)
+}
+